[handlers] separate handlers into unique files
